voice_agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.14.0
google-auth>=2.28.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
PyJWT>=2.8.0